Create README.md
Tom received a BA from Wabash College (1966) and served in the US Army from 1966-1969. He received a PhD in Botany from the University of Washington (1973) then spent three years at UBC on a postdoctoral fellowship before joining the Washington Department of Natural Resources in 1976. At WDNR he researched the cultivation of seaweeds for the production of phycolloids and food, and developed and managed programs for management and inventory of seaweeds and seagrasses on state-owned aquatic lands. He served on the Nearshore Science Team for the Puget Sound Nearshore Ecosystem Restoration Project. He has taught in various universities and consulted overseas in seaweed aquaculture. Since his retirement in 2011, he is focusing on researching marine algal biodiversity, the role of kelp in marine ecosystems, on kelp restoration, and on teaching. He was lead biologist for an ARPA-E project to grow kelp for biomass. He helped in developing the Puget Sound Kelp Conservation and Recovery Plan, and acts in an advisory role for the Puget Sound Restoration Fund’s project for using kelp to mitigate for OA. He sits as an alternate on the Strategic Advisory Council for the OCNMS and is on the Science Advisory Committee for the NW Straits Initiative.